move the listing to pkg so that it could be used by any other program as an API
